+++ differt/tests/test_differt_core.py
@@ -4,3 +4,4 @@ import differt_core
 
 def test_same_version() -> None:
     assert differt.__version__ == differt_core.__version__
+    assert differt.__version_info__ == differt_core.__version_info__